+++ .github/workflows/ci_linux/python_deps.sh
@@ -7,8 +7,7 @@ version=$(python -c 'import sys; version=sys.version_info[:2]; print("{0}.{1}".f
 if [ $version != '3.12' ]; then
     pip install geometric
     pip install spglib
-    pip install dftd3
-    pip install dftd4
+    pip install pyscf-dispersion
 fi
 
 #cppe
