+++ .github/workflows/ci_macos/build_pyscf.sh
@@ -6,7 +6,7 @@ cd ./pyscf/lib
 #curl -L https://github.com/fishjojo/pyscf-deps/raw/master/pyscf-1.7.5-deps-macos-10.14.tar.gz | tar xzf -
 mkdir build; cd build
 #cmake -DBUILD_LIBXC=OFF -DBUILD_XCFUN=OFF ..
-cmake -DCMAKE_Fortran_COMPILER=/usr/local/bin/gfortran ..
+cmake ..
 make -j4
 cd ..
 rm -Rf build
