+++ .github/workflows/ci_macos/deps_apt.sh
@@ -1,2 +1,1 @@
 #!/usr/bin/env bash
-brew install gfortran
