+++ pyscf/dft/test/test_h2o.py
@@ -18,6 +18,10 @@ import numpy
 from pyscf import gto
 from pyscf import lib
 from pyscf import dft
+try:
+    from pyscf.dispersion import dftd3, dftd4
+except ImportError:
+    dftd3 = dftd4 = None
 
 def setUpModule():
     global h2o, h2osym, h2o_cation, h2osym_cation
@@ -489,6 +493,7 @@ class KnownValues(unittest.TestCase):
         method.nlcgrids.atom_grid = {"H": (40, 110), "O": (40, 110),}
         self.assertAlmostEqual(method.scf(), -76.352381513158718, 8)
 
+    @unittest.skipIf(dftd3 is None, "requires the dftd3 library")
     def test_dft_parser(self):
         from pyscf.scf import dispersion
         method = dft.RKS(h2o, xc='wb97m-d3bj')
@@ -540,6 +545,7 @@ class KnownValues(unittest.TestCase):
         mf2.kernel()
         self.assertAlmostEqual(mf1.e_tot, -76.36649222362115, 9)
 
+    @unittest.skipIf(dftd3 is None, "requires the dftd3 library")
     def test_dispersion(self):
         mf = dft.RKS(h2o)
         mf.xc = 'B3LYP'
