+++ pyscf/grad/dispersion.py
@@ -26,6 +26,11 @@ from pyscf.dft import dft_parser
 
 def get_dispersion(mf_grad, disp_version=None, with_3body=False):
     '''gradient of dispersion correction for RHF/RKS'''
+    try:
+        from pyscf.dispersion import dftd3, dftd4
+    except ImportError:
+        print('dftd3 and dftd4 not available. Install them with `pip install pyscf-dispersion`')
+        raise
     mf = mf_grad.base
     mol = mf.mol
     if isinstance(mf, KohnShamDFT):
@@ -49,13 +54,11 @@ def get_dispersion(mf_grad, disp_version=None, with_3body=False):
         with_3body = mf.disp_with_3body
 
     if disp_version[:2].upper() == 'D3':
-        from pyscf.lib import dftd3
         d3_model = dftd3.DFTD3Dispersion(mol, xc=method, version=disp_version, atm=with_3body)
         res = d3_model.get_dispersion(grad=True)
         g_d3 = res.get('gradient')
         return g_d3
     elif disp_version[:2].upper() == 'D4':
-        from pyscf.lib import dftd4
         d4_model = dftd4.DFTD4Dispersion(mol, xc=method, atm=with_3body)
         res = d4_model.get_dispersion(grad=True)
         g_d4 = res.get('gradient')
