+++ pyscf/grad/test/test_rhf.py
@@ -18,6 +18,11 @@ import numpy
 from pyscf import gto, scf, lib
 from pyscf import grad
 
+try:
+    from pyscf.dispersion import dftd3, dftd4
+except ImportError:
+    dftd3 = dftd4 = None
+
 def setUpModule():
     global mol
     mol = gto.Mole()
@@ -71,6 +76,7 @@ class KnownValues(unittest.TestCase):
         e2 = mfs('O  0.  0.  0.001; H  0.  -0.757  0.587; H  0.  0.757   0.587')
         self.assertAlmostEqual(g[0,2], (e2-e1)/0.002*lib.param.BOHR, 5)
 
+    @unittest.skipIf(dftd3 is None, "requires the dftd3 library")
     def test_rhf_d3_grad(self):
         mf = scf.RHF(mol)
         mf.disp = 'd3bj'
@@ -82,6 +88,7 @@ class KnownValues(unittest.TestCase):
         e2 = mf_scan('O  0.  0.  0.001; H  0.  -0.757  0.587; H  0.  0.757   0.587')
         self.assertAlmostEqual((e2-e1)/0.002*lib.param.BOHR, g[0,2], 5)
 
+    @unittest.skipIf(dftd4 is None, "requires the dftd4 library")
     def test_rhf_d4_grad(self):
         mf = scf.RHF(mol)
         mf.disp = 'd4'
