+++ pyscf/hessian/test/test_rks.py
@@ -17,6 +17,10 @@ import unittest
 import numpy
 from pyscf import gto, dft, lib
 from pyscf import grad, hessian
+try:
+    from pyscf.dispersion import dftd3, dftd4
+except ImportError:
+    dftd3 = dftd4 = None
 
 def setUpModule():
     global mol, h4
@@ -118,7 +122,8 @@ class KnownValues(unittest.TestCase):
         #FIXME: errors seems too big
         self.assertAlmostEqual(abs(hess[0,:,2] - (e1-e2)/2e-4*lib.param.BOHR).max(), 0, 3)
 
-    def test_finite_diff_b3lyp_d3_hess(self):
+    @unittest.skipIf(dftd3 is None, "requires the dftd3 library")
+    def test_finite_diff_b3lyp_d3_hess_high_cost(self):
         mf = dft.RKS(mol)
         mf.conv_tol = 1e-14
         mf.xc = 'b3lyp'
@@ -133,6 +138,7 @@ class KnownValues(unittest.TestCase):
         #FIXME: errors seems too big
         self.assertAlmostEqual(abs(hess[0,:,2] - (e1-e2)/2e-4*lib.param.BOHR).max(), 0, 3)
 
+    @unittest.skipIf(dftd4 is None, "requires the dftd4 library")
     def test_finite_diff_b3lyp_d4_hess(self):
         mf = dft.RKS(mol)
         mf.conv_tol = 1e-14
