+++ pyscf/scf/dispersion.py
@@ -24,6 +24,11 @@ from pyscf.dft.rks import KohnShamDFT
 from pyscf.dft import dft_parser
 
 def get_dispersion(mf, disp_version=None):
+    try:
+        from pyscf.dispersion import dftd3, dftd4
+    except ImportError:
+        print('dftd3 and dftd4 not available. Install them with `pip install pyscf-dispersion`')
+        raise
     mol = mf.mol
     if isinstance(mf, KohnShamDFT):
         method = mf.xc
@@ -46,7 +51,6 @@ def get_dispersion(mf, disp_version=None):
 
     # for dftd3
     if disp_version[:2].upper() == 'D3':
-        from pyscf.lib import dftd3
         d3_model = dftd3.DFTD3Dispersion(mol, xc=method, version=disp_version, atm=with_3body)
         res = d3_model.get_dispersion()
         e_d3 = res.get('energy')
@@ -55,7 +59,6 @@ def get_dispersion(mf, disp_version=None):
 
     # for dftd4
     elif disp_version[:2].upper() == 'D4':
-        from pyscf.lib import dftd4
         d4_model = dftd4.DFTD4Dispersion(mol, xc=method, atm=with_3body)
         res = d4_model.get_dispersion()
         e_d4 = res.get('energy')
